+++ app/Livewire/BranchDashboard/Branches/DeleteBranch.php
@@ -217,7 +217,7 @@ class DeleteBranch extends BaseComponent
 
     public function render()
     {
-        $rows = $this->getFilteredQuery()->orderBy('created_at', 'DESC')->paginate($this->quantity ?? 10);
+        $rows = $this->getFilteredQuery()->orderBy('created_at', 'DESC')->paginate((int) ($this->quantity ?? 10));
         $users = User::all();
 
         return view('livewire.branch-dashboard.branches.delete-branch', [
