+++ app/Livewire/BranchDashboard/Production/ProductTypes.php
@@ -172,7 +172,7 @@ class ProductTypes extends BaseComponent
         }
 
         $rows = Cache::remember($this->getCacheKey(), now()->addMinutes($this->cacheTtlMinutes), function () {
-            return $this->getFilteredQuery()->paginate($this->quantity ?? 10);
+            return $this->getFilteredQuery()->paginate((int) ($this->quantity ?? 10));
         });
 
         // Add sequential numbers to the rows
